test: ProjectManagerTests 파일 추가 #3
