+++ ProjectManager/ProjectManager/ViewModel/TaskListViewModel.swift
@@ -1,16 +1,5 @@
 import Foundation
 
-enum TaskManagerError: Error {
-    case taskNotFound
-    
-    var description: String {
-        switch self {
-        case .taskNotFound:
-            return "존재하지 않는 Task입니다."
-        }
-    }
-}
-
 protocol TaskListViewModelProtocol: AnyObject {
     var todoTasksObservable: MockObservable<[Task]>! { get }
     var doingTasksObservable: MockObservable<[Task]>! { get }
